requirements.txt
Django==6.0.3
gunicorn==25.3.0
whitenoise==6.12.0
python-dotenv